Refactor package details content
