+++ web/src/layout/package/__fixtures__/Details/1.json
