+++ web/src/layout/package/__fixtures__/Details/10.json
