+++ web/src/layout/package/__fixtures__/Details/11.json
