+++ web/src/layout/package/__fixtures__/Details/13.json
