+++ web/src/layout/package/__fixtures__/Details/14.json
