+++ web/src/layout/package/__fixtures__/Details/16.json
