+++ web/src/layout/package/__fixtures__/Details/17.json
@@ -5,18 +5,17 @@
   "description": "desc",
   "logoImageId": "imageId",
   "appVersion": "1.0.0",
-  "userAlias": null,
   "normalizedName": "pr",
   "deprecated": false,
-  "isOperator": false,
   "keywords": ["key1", "key2"],
+  "license": "MIT",
+  "createdAt": 0,
+  "signed": false,
   "repository": {
-    "repositoryid": "id",
-    "kind": 0,
-    "name": "incubator",
-    "displayName": null,
-    "organizationName": "helm",
-    "organizationDisplayName": "Helm",
-    "userAlias": null
+    "kind": 1,
+    "name": "repo",
+    "displayName": "Repo",
+    "url": "http://repo.test",
+    "userAlias": "user"
   }
 }
