+++ web/src/layout/package/__fixtures__/Details/2.json
@@ -1,0 +1,22 @@
+{
+  "packageId": "id",
+  "name": "test",
+  "displayName": "Pretty name",
+  "description": "desc",
+  "logoImageId": "imageId",
+  "appVersion": "1.0.0",
+  "normalizedName": "pr",
+  "maintainers": [{ "email": "main@tainer.com", "name": "maintainerName" }],
+  "deprecated": false,
+  "repository": {
+    "kind": 0,
+    "name": "stable",
+    "displayName": null,
+    "url": "repoUrl",
+    "userAlias": "user"
+  },
+  "license": "MIT",
+  "createdAt": 1,
+  "signed": false,
+  "capabilities": "basic install"
+}
