+++ web/src/layout/package/__fixtures__/Details/4.json
