+++ web/src/layout/package/__fixtures__/Details/6.json
