+++ web/src/layout/package/__fixtures__/Details/7.json
