+++ web/src/layout/package/__fixtures__/Details/8.json
