+++ web/src/layout/package/__fixtures__/Details/9.json
