--- web/src/layout/package/__fixtures__/details/8.json
@@ -1,19 +1,0 @@
-{
-  "packageId": "id",
-  "name": "chartmuseum",
-  "displayName": null,
-  "description": null,
-  "homeUrl": null,
-  "logoImageId": null,
-  "appVersion": "1.1.1",
-  "maintainers": null,
-  "repository": {
-    "repositoryid": "id",
-    "kind": 0,
-    "name": "incubator",
-    "displayName": null,
-    "organizationName": "helm",
-    "organizationDisplayName": "Helm",
-    "userAlias": null
-  }
-}
